feat(organize-files: client/app , components/drawer-navabar /copyright /index)drawer nav and copyright are their own components, app.js is now simplified.
