+++ client/components/drawer-navbar.js
@@ -1,0 +1,239 @@
+import React from 'react'
+import clsx from 'clsx'
+import {makeStyles} from '@material-ui/core/styles'
+import CssBaseline from '@material-ui/core/CssBaseline'
+import Drawer from '@material-ui/core/Drawer'
+import AppBar from '@material-ui/core/AppBar'
+import Toolbar from '@material-ui/core/Toolbar'
+import Typography from '@material-ui/core/Typography'
+import Divider from '@material-ui/core/Divider'
+import IconButton from '@material-ui/core/IconButton'
+import MenuOpenIcon from '@material-ui/icons/MenuOpen'
+import MenuIcon from '@material-ui/icons/Menu'
+import HomeIcon from '@material-ui/icons/Home'
+import PersonIcon from '@material-ui/icons/Person'
+import TextFieldsIcon from '@material-ui/icons/TextFields'
+import PropTypes from 'prop-types'
+import {connect} from 'react-redux'
+import {logout} from '../store'
+import ContactMailIcon from '@material-ui/icons/ContactMail'
+import ListItemIcon from '@material-ui/core/ListItemIcon'
+import ListItemText from '@material-ui/core/ListItemText'
+import ListItem from '@material-ui/core/ListItem'
+import {Link as RLink} from 'react-router-dom'
+import MeetingRoomIcon from '@material-ui/icons/MeetingRoom'
+import ExitToAppIcon from '@material-ui/icons/ExitToApp'
+
+const drawerWidth = 240
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    display: 'flex'
+  },
+  toolbar: {
+    paddingRight: 20, // keep right padding when drawer closed
+    color: 'white'
+  },
+  toolbarIcon: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'flex-end',
+    padding: '0 8px',
+    ...theme.mixins.toolbar
+  },
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+    transition: theme.transitions.create(['width', 'margin'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    })
+  },
+  appBarShift: {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(['width', 'margin'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  menuButton: {
+    marginRight: 36
+  },
+  menuButtonHidden: {
+    display: 'none'
+  },
+  title: {
+    flexGrow: 1
+  },
+  drawerPaper: {
+    position: 'relative',
+    whiteSpace: 'nowrap',
+    width: drawerWidth,
+    transition: theme.transitions.create('width', {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  drawerPaperClose: {
+    overflowX: 'hidden',
+    transition: theme.transitions.create('width', {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    }),
+    width: theme.spacing(7),
+    [theme.breakpoints.up('sm')]: {
+      width: theme.spacing(9)
+    }
+  },
+  appBarSpacer: theme.mixins.toolbar,
+  content: {
+    flexGrow: 1,
+    height: '100vh',
+    overflow: 'auto'
+  },
+  container: {
+    paddingTop: theme.spacing(4),
+    paddingBottom: theme.spacing(4)
+  },
+  paper: {
+    padding: theme.spacing(2),
+    display: 'flex',
+    overflow: 'auto',
+    flexDirection: 'column'
+  },
+  fixedHeight: {
+    height: 240
+  }
+}))
+
+const DrawerNavbar = ({handleClick, isLoggedIn}) => {
+  const classes = useStyles()
+
+  const [open, setOpen] = React.useState(false)
+
+  const handleDrawerOpen = () => {
+    setOpen(true)
+  }
+  const handleDrawerClose = () => {
+    setOpen(false)
+  }
+  //const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight)
+
+  return (
+    <div className={classes.root}>
+      <CssBaseline />
+      <AppBar
+        color="default"
+        position="absolute"
+        className={clsx(classes.appBar, open && classes.appBarShift)}
+      >
+        <Toolbar className={classes.toolbar}>
+          <IconButton
+            edge="start"
+            color="inherit"
+            aria-label="open drawer"
+            onClick={handleDrawerOpen}
+            className={clsx(
+              classes.menuButton,
+              open && classes.menuButtonHidden
+            )}
+          >
+            <MenuIcon color="action" />
+          </IconButton>
+          <img src="/images/boss.png" alt="logo" width="250" />
+        </Toolbar>
+      </AppBar>
+      <Drawer
+        variant="permanent"
+        classes={{
+          paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose)
+        }}
+        open={open}
+      >
+        <div className={classes.toolbarIcon}>
+          <IconButton onClick={handleDrawerClose}>
+            <MenuOpenIcon />
+          </IconButton>
+        </div>
+        <Divider />
+        <div>
+          <ListItem button component={RLink} to="/home">
+            <ListItemIcon>
+              <HomeIcon href="/home" />
+            </ListItemIcon>
+            <ListItemText primary="Home" />
+          </ListItem>
+          <Divider />
+          <ListItem button component={RLink} to="/texteditor">
+            <ListItemIcon>
+              <TextFieldsIcon />
+            </ListItemIcon>
+            <ListItemText primary="Text Editor" />
+          </ListItem>
+          <Divider />
+          {isLoggedIn ? (
+            <div>
+              <Divider />
+              <ListItem button component={RLink} to="/myaccount">
+                <ListItemIcon>
+                  <ContactMailIcon />
+                </ListItemIcon>
+                <ListItemText primary="My Account" />
+              </ListItem>
+              <Divider />
+              <ListItem
+                button
+                component={RLink}
+                to="/login"
+                onClick={handleClick}
+              >
+                <ListItemIcon>
+                  <MeetingRoomIcon />
+                </ListItemIcon>
+                <ListItemText primary="Log Out" />
+              </ListItem>
+              <Divider />
+            </div>
+          ) : (
+            <div>
+              <ListItem button component={RLink} to="/login">
+                <ListItemIcon>
+                  <ExitToAppIcon />
+                </ListItemIcon>
+                <ListItemText primary="Log In" />
+              </ListItem>
+              <Divider />
+              <ListItem button component={RLink} to="/signup">
+                <ListItemIcon>
+                  <PersonIcon />
+                </ListItemIcon>
+                <ListItemText primary="Sign Up" />
+              </ListItem>
+              <Divider />
+            </div>
+          )}
+        </div>
+      </Drawer>
+    </div>
+  )
+}
+
+const mapState = state => {
+  return {
+    isLoggedIn: !!state.user.id
+  }
+}
+
+const mapDispatch = dispatch => {
+  return {
+    handleClick() {
+      dispatch(logout())
+    }
+  }
+}
+
+DrawerNavbar.propTypes = {
+  handleClick: PropTypes.func.isRequired,
+  isLoggedIn: PropTypes.bool.isRequired
+}
+export default connect(mapState, mapDispatch)(DrawerNavbar)
