+++ client/components/index.js
@@ -6,3 +6,5 @@ export {default as TextEditor} from './text-editor'
 export {default as Home} from './home'
 export {default as MyAccount} from './my-account'
 export {default as Drafts} from './drafts'
+export {default as DrawerNavbar} from './drawer-navbar'
+export {default as Copyright} from './copyright'
