+++ public/style.css
@@ -39,20 +39,20 @@ mark {
 
 #home-image-wrapper {
   position: relative;
-  width: 100%;
+  max-width: 100%;
   margin-left: -5.35%;
   margin-top: -2.7%;
 }
 
-.home-main-image {
+/* .home-main-img {
   position: relative;
-  /* width: 100%;
-  height: 100%; */
+  width: 100%;
+  height: 100%;
   max-width: 100%;
   height: auto;
   width: auto\9;
   background-size: cover;
-}
+} */
 
 #home-text {
   position: absolute;
